workshops/dlt/tfl-pipeline/pyproject.toml
[project]
name = "tfl-pipeline"
version = "0.1.0"
description = "dlt pipeline for Transport for London Unified API"
requires-python = ">=3.11"
dependencies = [
    "dlt[duckdb]>=1.21.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "dlt[workspace]>=1.21.0",
    "ibis-framework[duckdb]>=12.0.0",
    "marimo>=0.19.9",
    "altair>=6.0.0",
]
